restructured sudoku game dir
