+++ src/sudoku/View/style/sudokuScreenStyleSheet.css
